requirements.txt
openai>=1.0.0
pydantic>=2.0.0
click>=8.0.0
rich>=13.0.0
platformdirs>=3.0.0
tomli>=2.0.0
tiktoken>=0.5.0
httpx>=0.24.0
duckduckgo-search>=8.0.0
python-dotenv>=1.0.0
PyGithub>=2.1.1
pytest>=7.0.0
python-telegram-bot>=20.0
